Opportunistically move a couple more compile-only tests over to the right module
